+++ components/icon-font/src/lib/fas-solid-pro.ts
@@ -105,5 +105,10 @@ export {
   faEye,
   faGear,
   faUnlock,
-  faBell
+  faBell,
+  fa1,
+  fa2,
+  fa3,
+  faPen,
+  faUserGroup
 } from "@fortawesome/pro-solid-svg-icons";